+++ src/lib.rs
@@ -223,7 +223,7 @@ mod de;
 mod ser;
 
 pub use reading::{read, Reader, Record, FieldIterator, ReadableRecord, NamedValue};
-pub use record::field::{FieldValue, Date, DateTime};
+pub use record::field::{FieldValue, Date, Time, DateTime};
 pub use record::{FieldInfo, FieldName, FieldFlags, FieldConversionError};
 pub use writing::{TableWriter, TableWriterBuilder, WritableRecord, FieldWriter};
 use std::fmt::{Display, Formatter};
@@ -251,6 +251,7 @@ pub enum Error {
     /// was expected
     NotEnoughFields,
     BadFieldType{expected: FieldType, got: FieldType, field_name: String},
+    IncompatibleType,
     Message(String),
 }
 
@@ -302,6 +303,7 @@ impl std::error::Error for Error {
             Error::BadFieldType { expected: _, got: _, field_name: _ } => {
                 "The Given type does not match the expected field type"
             }
+            Error::IncompatibleType => {"type is not compatible"}
         }
     }
 }