+++ src/record/mod.rs
@@ -1,4 +1,4 @@
-use std::convert::TryFrom;
+use std::convert::{TryFrom, TryInto};
 use std::io::{Read, Write};
 
 use byteorder::{ReadBytesExt, WriteBytesExt};
@@ -6,7 +6,8 @@ use byteorder::{ReadBytesExt, WriteBytesExt};
 pub mod field;
 
 use ::{Error, FieldValue};
-use record::field::{FieldType, Date};
+use record::field::{FieldType, Date, DateTime};
+use Error::IncompatibleType;
 
 
 const DELETION_FLAG_NAME: &'static str = "DeletionFlag";
@@ -223,7 +224,7 @@ macro_rules! impl_try_from_field_value_for_ {
 }
 
 impl_try_from_field_value_for_!(FieldValue::Numeric => Option<f64>);
-impl_try_from_field_value_for_!(FieldValue::Numeric(Some(v)) => f64);
+//impl_try_from_field_value_for_!(FieldValue::Numeric(Some(v)) => f64);
 
 impl_try_from_field_value_for_!(FieldValue::Float => Option<f32>);
 impl_try_from_field_value_for_!(FieldValue::Float(Some(v)) => f32);
@@ -236,6 +237,26 @@ impl_try_from_field_value_for_!(FieldValue::Character(Some(string)) => String);
 
 impl_try_from_field_value_for_!(FieldValue::Logical => Option<bool>);
 impl_try_from_field_value_for_!(FieldValue::Logical(Some(b)) => bool);
+
+impl_try_from_field_value_for_!(FieldValue::Integer => i32);
+
+impl TryFrom<FieldValue> for f64 {
+    type Error = Error;
+
+    fn try_from(value: FieldValue) -> Result<Self, Self::Error> {
+        match value {
+            FieldValue::Numeric(Some(v)) => Ok(v),
+            FieldValue::Numeric(None) => Err(FieldConversionError::NoneValue.into()),
+            FieldValue::Currency(c) => Ok(c),
+            FieldValue::Double(d) => Ok(d),
+            _ => Err(IncompatibleType)
+        }
+    }
+}
+
+
+// Fox Pro types
+impl_try_from_field_value_for_!(FieldValue::DateTime => DateTime);
 
 macro_rules! impl_from_type_for_field_value (
     ($t:ty => FieldValue::$variant:ident) => {
@@ -268,6 +289,9 @@ impl_from_type_for_field_value!(bool => FieldValue::Logical(Some(v)));
 
 impl_from_type_for_field_value!(Option<Date> => FieldValue::Date);
 impl_from_type_for_field_value!(Date => FieldValue::Date(Some(v)));
+
+// Fox Pro types
+impl_from_type_for_field_value!(DateTime => FieldValue::DateTime);
 
 #[cfg(test)]
 mod test {